Updates dependencies and adapts to datagrid changes
Updates the minimum PHP version to 8.0.

Adapts the QueryObjectDataSource to breaking changes in the contributte/datagrid package.
Specifically updates namespaces and usage of DateTimeHelper.
Also updates the QueryObject usage to use the interface instead of the implementation and removes unused property and redundant phpDoc.

This ensures compatibility with newer versions of the datagrid library.